Removed PlaceBetAndMove from core and moved its methods down. It was only used the once, and I didn't see any other use cases for it.
