+++ crapssim/strategy/core.py
@@ -268,128 +268,6 @@ class RemoveByType(RemoveIfTrue):
         super().__init__(lambda b, p: isinstance(b, bet_type))
 
 
-class PlaceBetAndMove(Strategy):
-    """Strategy that makes Place bets and then moves the bet to other Places if a PassLine,
-    DontPass, Come or DontCOme bet gets moved to a bet with the same number."""
-    def __init__(self, starting_bets: list[Place],
-                 check_bets: list[PassLine | DontPass | Come | DontCome],
-                 bet_movements: dict[Place, Place | None]):
-        """Makes the starting place bets in starting_bets and then if one of the check_bets gets
-        moved to the same point as one of the place bets, the bet gets moved to a different bet
-        depending on the bet_movements dictionary.
-
-        Parameters
-        ----------
-        starting_bets
-            Starting bets placed on the table.
-        check_bets
-            If one of these bets ended up having the same point as one of the place bets,
-            move to a different bet.
-        bet_movements
-            If the place bet is at a point of a check bet return the bet to replace it with
-            or None to remove it.
-        """
-
-        self.starting_bets = starting_bets
-        self.check_bets = check_bets
-        self.bet_movements = bet_movements
-
-    def check_bets_on_table(self, player: 'Player') -> list[PassLine | DontPass | Come | DontCome]:
-        """Returns any bets the player has on the table that are in check_bets.
-
-        Parameters
-        ----------
-        player
-            The player to check the bets for.
-
-        Returns
-        -------
-        list[AllowsOdds]
-            A list of all the check bets that are on the table.
-        """
-        return [x for x in player.bets_on_table if x in self.check_bets if
-                isinstance(x, (PassLine, DontPass, Come, DontCome))]
-
-    def check_numbers(self, player: 'Player') -> list[int]:
-        """Returns the points of all the check bets that are currently on the table.
-
-        Parameters
-        ----------
-        player
-            The player to get the check bets points from.
-
-        Returns
-        -------
-        list[int]
-            A list of points of bets that are check_bets the player has on the table.
-        """
-        check_numbers = []
-        for bet in self.check_bets:
-            if bet in player.bets_on_table:
-                check_numbers += bet.get_winning_numbers(player.table)
-        return check_numbers
-
-    def place_starting_bets(self, player: 'Player') -> None:
-        """Place the initial place bets.
-
-        Parameters
-        ----------
-        player
-            The player to place the bets for.
-        """
-        for bet in self.starting_bets:
-            if bet not in player.bets_on_table and player.table.point.status != "Off":
-                player.add_bet(bet)
-
-    def bets_to_move(self, player: 'Player') -> list[Place]:
-        """A list of the bets that need to bet moved to a different bet.
-
-        Parameters
-        ----------
-        player
-            The player to place the bets for.
-
-        Returns
-        -------
-        list[Place]
-            A list of the bets that need to be moved to a different bet.
-        """
-        return [x for x in self.bet_movements if x.winning_numbers[0] in
-                self.check_numbers(player) and x in player.bets_on_table]
-
-    def move_bets(self, player: 'Player') -> None:
-        """Move any bets that need to be moved to a different bet as determined by bet_movements.
-
-        Parameters
-        ----------
-        player
-            The player to move the bets for.
-        """
-        while len(self.bets_to_move(player)) > 0:
-            old_bet = self.bets_to_move(player)[0]
-            new_bet = self.bet_movements[old_bet]
-            while new_bet in player.bets_on_table:
-                new_bet = self.bet_movements[new_bet]
-            player.remove_bet(old_bet)
-            if new_bet is not None:
-                player.add_bet(new_bet)
-
-    def update_bets(self, player: 'Player') -> None:
-        """Place the initial bets and move them to the desired location.
-
-        Parameters
-        ----------
-        player
-            The player to move the bets for.
-        """
-        self.place_starting_bets(player)
-        self.move_bets(player)
-
-    def __repr__(self) -> str:
-        return f'{self.__class__.__name__}(starting_bets={self.starting_bets}, ' \
-               f'check_bets={self.check_bets}, bet_movements={self.bet_movements})'
-
-
 class OddsStrategy(Strategy):
     """Strategy that takes an AllowsOdds object and places Odds on it given either a multiplier,
     or a dictionary of points and multipliers."""
